+++ Package.swift
@@ -2,8 +2,8 @@
 
 import PackageDescription
 
-let tag = "v0.1.0-rc41"
-let checksum = "fb3d8b3747db64f19b24290df814f99fb2a4b190b458d44ee949459b1d30be39"
+let tag = "v0.1.0-rc42"
+let checksum = "1223552e563ab1c6fc1fd0dc3e830e412c7a29cabf9b7bc91384ef9651ea4766"
 let url = "https://github.com/pubky/paykit-rs/releases/download/\(tag)/Paykit.xcframework.zip"
 
 let package = Package(