+++ paykit-ffi/bindings/ios/paykit.swift
@@ -11384,13 +11384,9 @@ public struct PubkyClientConfig {
      */
     public var requestTimeoutSecs: UInt64
     /**
-     * Pubky network environment used by the client.
+     * Host running local testnet services, or `None` to use the public Pubky network.
      */
-    public var environment: PubkyClientEnvironment
-    /**
-     * DNS hostname or IPv4 address running local testnet services, or `None` for localhost.
-     */
-    public var testnetHost: String?
+    public var localTestnetHost: String?
 
     // Default memberwise initializers are never public by default, so we
     // declare one manually.
@@ -11399,14 +11395,10 @@ public struct PubkyClientConfig {
          * Request timeout for Pubky HTTP operations in seconds.
          */requestTimeoutSecs: UInt64,
         /**
-         * Pubky network environment used by the client.
-         */environment: PubkyClientEnvironment,
-        /**
-         * DNS hostname or IPv4 address running local testnet services, or `None` for localhost.
-         */testnetHost: String?) {
+         * Host running local testnet services, or `None` to use the public Pubky network.
+         */localTestnetHost: String?) {
         self.requestTimeoutSecs = requestTimeoutSecs
-        self.environment = environment
-        self.testnetHost = testnetHost
+        self.localTestnetHost = localTestnetHost
     }
 }
 
@@ -11420,10 +11412,7 @@ extension PubkyClientConfig: Equatable, Hashable {
         if lhs.requestTimeoutSecs != rhs.requestTimeoutSecs {
             return false
         }
-        if lhs.environment != rhs.environment {
-            return false
-        }
-        if lhs.testnetHost != rhs.testnetHost {
+        if lhs.localTestnetHost != rhs.localTestnetHost {
             return false
         }
         return true
@@ -11431,8 +11420,7 @@ extension PubkyClientConfig: Equatable, Hashable {
 
     public func hash(into hasher: inout Hasher) {
         hasher.combine(requestTimeoutSecs)
-        hasher.combine(environment)
-        hasher.combine(testnetHost)
+        hasher.combine(localTestnetHost)
     }
 }
 
@@ -11448,15 +11436,13 @@ public struct FfiConverterTypePubkyClientConfig: FfiConverterRustBuffer {
         return
             try PubkyClientConfig(
                 requestTimeoutSecs: FfiConverterUInt64.read(from: &buf),
-                environment: FfiConverterTypePubkyClientEnvironment.read(from: &buf),
-                testnetHost: FfiConverterOptionString.read(from: &buf)
+                localTestnetHost: FfiConverterOptionString.read(from: &buf)
         )
     }
 
     public static func write(_ value: PubkyClientConfig, into buf: inout [UInt8]) {
         FfiConverterUInt64.write(value.requestTimeoutSecs, into: &buf)
-        FfiConverterTypePubkyClientEnvironment.write(value.environment, into: &buf)
-        FfiConverterOptionString.write(value.testnetHost, into: &buf)
+        FfiConverterOptionString.write(value.localTestnetHost, into: &buf)
     }
 }
 
@@ -15349,97 +15335,6 @@ public func FfiConverterTypePubkyAuthRequestKind_lower(_ value: PubkyAuthRequest
 extension PubkyAuthRequestKind: Equatable, Hashable {}
 
 extension PubkyAuthRequestKind: Codable {}
-
-
-
-
-
-
-// Note that we don't yet support `indirect` for enums.
-// See https://github.com/mozilla/uniffi-rs/issues/396 for further discussion.
-/**
- * Pubky network environment used by binding-layer clients.
- */
-
-public enum PubkyClientEnvironment {
-
-    /**
-     * Use the public Pubky network.
-     */
-    case production
-    /**
-     * Use standard Pubky testnet ports, on localhost unless a host is configured.
-     */
-    case localTestnet
-    /**
-     * SDK returned a value this binding version does not understand.
-     */
-    case unknown
-}
-
-
-#if compiler(>=6)
-extension PubkyClientEnvironment: Sendable {}
-#endif
-
-#if swift(>=5.8)
-@_documentation(visibility: private)
-#endif
-public struct FfiConverterTypePubkyClientEnvironment: FfiConverterRustBuffer {
-    typealias SwiftType = PubkyClientEnvironment
-
-    public static func read(from buf: inout (data: Data, offset: Data.Index)) throws -> PubkyClientEnvironment {
-        let variant: Int32 = try readInt(&buf)
-        switch variant {
-
-        case 1: return .production
-
-        case 2: return .localTestnet
-
-        case 3: return .unknown
-
-        default: throw UniffiInternalError.unexpectedEnumCase
-        }
-    }
-
-    public static func write(_ value: PubkyClientEnvironment, into buf: inout [UInt8]) {
-        switch value {
-
-
-        case .production:
-            writeInt(&buf, Int32(1))
-
-
-        case .localTestnet:
-            writeInt(&buf, Int32(2))
-
-
-        case .unknown:
-            writeInt(&buf, Int32(3))
-
-        }
-    }
-}
-
-
-#if swift(>=5.8)
-@_documentation(visibility: private)
-#endif
-public func FfiConverterTypePubkyClientEnvironment_lift(_ buf: RustBuffer) throws -> PubkyClientEnvironment {
-    return try FfiConverterTypePubkyClientEnvironment.lift(buf)
-}
-
-#if swift(>=5.8)
-@_documentation(visibility: private)
-#endif
-public func FfiConverterTypePubkyClientEnvironment_lower(_ value: PubkyClientEnvironment) -> RustBuffer {
-    return FfiConverterTypePubkyClientEnvironment.lower(value)
-}
-
-
-extension PubkyClientEnvironment: Equatable, Hashable {}
-
-extension PubkyClientEnvironment: Codable {}